requirements.txt

pandas
SQLAlchemy
pyodbc
mysql-connector-python

python-dotenv == 1.0.0
openai
fastapi 
uvicorn
